refactor: fix files name to ReactJs convention
