+++ src/components/TimeTable.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import Block from "./block.js";
+import Block from "./Block.js";
 
 // table part contain Block
 class TimeTable extends Component {
